requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-google
livekit-plugins-silero
livekit-plugins-noise-cancellation
mem0ai
google-api-python-client
langchain_community
requests
python-dotenv
